Value input based on field type + decimal -> double + mapping to dynamic playlist map in arrangement

+++ MapMaven.Core/Models/Data/ScoreSaber/RankedMap.cs
@@ -11,24 +11,24 @@ namespace MapMaven.Core.Models.Data.ScoreSaber
         public string Name { get; set; }
         public string Artist { get; set; }
         public string Mapper { get; set; }
-        public decimal Bpm { get; set; }
+        public double Bpm { get; set; }
         [JsonPropertyName("diff")]
         public string Difficulty { get; set; }
         public int Scores { get; set; }
-        public decimal Stars { get; set; }
-        public decimal PP { get; set; }
+        public double Stars { get; set; }
+        public double PP { get; set; }
         [JsonPropertyName("beatSaverKey")]
         public string Key { get; set; }
         public int Downloads { get; set; }
         public int Upvotes { get; set; }
         public int Downvotes { get; set; }
-        public decimal Rating { get; set; }
+        public double Rating { get; set; }
         public string Download { get; set; }
-        public decimal Duration { get; set; }
-        public decimal? DurationSeconds { get; set; }
+        public double Duration { get; set; }
+        public double? DurationSeconds { get; set; }
         public int NoteCount { get; set; }
         [JsonPropertyName("njs")]
-        public decimal NoteJumpSpeed { get; set; }
+        public double NoteJumpSpeed { get; set; }
         public int RecentScores { get; set; }
 
         public Map ToMap()
